✨ Entity 초기 설계
