--- src/main/java/com/mycom/backenddaengplace/common/domain/member/entity/type/Gender.java
@@ -1,7 +1,0 @@
-package com.mycom.backenddaengplace.common.domain.member.entity.type;
-
-public enum Gender {
-    MALE,
-    FEMALE,
-    OTHER
-}
