--- src/main/java/com/mycom/backenddaengplace/common/domain/member/entity/type/SocialType.java
@@ -1,5 +1,0 @@
-package com.mycom.backenddaengplace.common.domain.member.entity.type;
-
-public enum SocialType {
-    KAKAO, GOOGLE,
-}
