--- src/main/java/com/mycom/backenddaengplace/common/domain/pet/entity/BreedType.java
@@ -1,25 +1,0 @@
-package com.mycom.backenddaengplace.common.domain.pet.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Table(name = "breed_type")
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class BreedType {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "breed_type_id")
-    private Long id;
-
-    private String breedType;
-
-}
