--- src/main/java/com/mycom/backenddaengplace/common/domain/place/entity/Address.java
@@ -1,29 +1,0 @@
-package com.mycom.backenddaengplace.common.domain.place.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "address")
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Address {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "address_id")
-    private Long id;
-
-    @Column(name = "road_address", nullable = false)
-    private String roadAddress;
-
-    @Column(name = "detail_address")
-    private String detailAddress;
-
-    @Column(name = "zipcode")
-    private String zipcode;
-
-}
