--- src/main/java/com/mycom/backenddaengplace/common/domain/place/entity/Location.java
@@ -1,28 +1,0 @@
-package com.mycom.backenddaengplace.common.domain.place.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "location")
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Location {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long locationId;
-
-    @ManyToOne
-    @JoinColumn(name = "address_id", nullable = false)
-    private Address address;
-
-    @Column(name = "latitude", precision = 10, scale = 7)
-    private Double latitude;
-
-    @Column(name = "longitude", precision = 11, scale = 7)
-    private Double longitude;
-}
