--- src/main/java/com/mycom/backenddaengplace/common/domain/place/entity/enums/Category.java
@@ -1,6 +1,0 @@
-package com.mycom.backenddaengplace.common.domain.place.entity.enums;
-
-public enum Category {
-    RESTAURANT, PARK, HOSPITAL, HOTEL
-}
-
