--- src/main/java/com/mycom/backenddaengplace/common/domain/place/entity/enums/WeatherType.java
@@ -1,5 +1,0 @@
-package com.mycom.backenddaengplace.common.domain.place.entity.enums;
-
-public enum WeatherType {
-    RAIN, SUNNY, CLOUDY
-}
