+++ src/main/java/com/mycom/backenddaengplace/member/domain/Gender.java
@@ -1,0 +1,7 @@
+package com.mycom.backenddaengplace.member.domain;
+
+public enum Gender {
+    MALE,
+    FEMALE,
+    OTHER
+}
