+++ src/main/java/com/mycom/backenddaengplace/member/domain/SocialType.java
@@ -1,0 +1,5 @@
+package com.mycom.backenddaengplace.member.domain;
+
+public enum SocialType {
+    KAKAO, GOOGLE,
+}
