+++ src/main/java/com/mycom/backenddaengplace/place/domain/Address.java
@@ -1,0 +1,29 @@
+package com.mycom.backenddaengplace.place.domain;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "address")
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Address {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "address_id")
+    private Long id;
+
+    @Column(name = "road_address", nullable = false)
+    private String roadAddress;
+
+    @Column(name = "detail_address")
+    private String detailAddress;
+
+    @Column(name = "zipcode")
+    private String zipcode;
+
+}
