+++ src/main/java/com/mycom/backenddaengplace/place/domain/Category.java
@@ -1,0 +1,6 @@
+package com.mycom.backenddaengplace.place.domain;
+
+public enum Category {
+    RESTAURANT, PARK, HOSPITAL, HOTEL
+}
+
