+++ src/main/java/com/mycom/backenddaengplace/place/domain/WeatherType.java
@@ -1,0 +1,5 @@
+package com.mycom.backenddaengplace.place.domain;
+
+public enum WeatherType {
+    RAIN, SUNNY, CLOUDY
+}
